+++ src/UrlDependencyResolver.js
@@ -26,8 +26,7 @@ const UrlDependencyResolver = {
    * @private
    */
   resolveInPaths(files, request) {
-    /** @type {FileSystem} fs */
-    const { fs } = this;
+    const fs = this.fs;
     let context, file, tmpFile;
     for (tmpFile of files) {
       context = path.dirname(tmpFile);
@@ -45,18 +44,17 @@ const UrlDependencyResolver = {
 
   /**
    * @param {{snapshot: {managedFiles: Set<string>, children: Set<{managedFiles: Set<string>}>}}} buildInfo
-   * @param {string} request The file to be resolved.
+   * @param {string} resource The file to be resolved.
    * @returns {null|{file: string, context: string}}
    * @private
    */
-  resolveInBuildInfo(buildInfo, request) {
-    const self = this;
+  resolveInBuildInfo(buildInfo, resource) {
     const { snapshot } = buildInfo;
     const { managedFiles, children } = snapshot;
     let result;
 
     if (managedFiles != null && managedFiles.size > 0) {
-      result = self.resolveInPaths(managedFiles, request);
+      result = this.resolveInPaths(managedFiles, resource);
       if (result != null) {
         return result;
       }
@@ -66,7 +64,7 @@ const UrlDependencyResolver = {
       for (let item of children) {
         const childrenManagedFiles = item.managedFiles;
         if (childrenManagedFiles != null && childrenManagedFiles.size > 0) {
-          result = self.resolveInPaths(childrenManagedFiles, request);
+          result = this.resolveInPaths(childrenManagedFiles, resource);
           if (result != null) {
             return result;
           }
@@ -83,28 +81,53 @@ const UrlDependencyResolver = {
    * @param {{}} resolveData The callback parameter for the hooks beforeResolve of NormalModuleFactory.
    */
   resolve(resolveData) {
-    const self = this;
-    const fs = self.fs;
-    const request = resolveData.request;
+    const fs = this.fs;
+    const [resource, query] = resolveData.request.split('?');
 
-    if (!fs.existsSync(path.resolve(resolveData.context, request))) {
+    if (!fs.existsSync(path.resolve(resolveData.context, resource))) {
       const dependency = resolveData.dependencies[0];
       const parentModule = this.moduleGraph.getParentModule(dependency);
       const buildInfo = parentModule.buildInfo || {};
       const snapshot = buildInfo.snapshot || {};
-      const issuers = snapshot.fileTimestamps || snapshot.fileTshs;
-      /** @type {string} closest issuer that can import the resource */
-      const closestIssuer = issuers != null && issuers.size > 0 ? Array.from(issuers.keys()).pop() : null;
-      let context = closestIssuer ? path.dirname(closestIssuer) : resolveData.context;
+      const dependencies = snapshot.fileTimestamps || snapshot.fileTshs;
+
+      let context = resolveData.context;
       let resolvedFile;
 
-      // 1. try to resolve relative path in context or in directory of the closest issuer
-      let tmpFile = path.resolve(context, request);
-      if (fs.existsSync(tmpFile)) {
-        resolvedFile = tmpFile;
+      // 1. try to resolve relative path in context or in dependency directories
+      if (dependencies != null && dependencies.size > 0) {
+        const files = dependencies.keys();
+        const cache = new Set();
+        let file;
+
+        for (file of files) {
+          if (fs.lstatSync(file).isFile()) {
+            let dir = path.dirname(file);
+            if (cache.has(dir)) {
+              // skip directory that is already checked and the resource was not resolved in this directory
+              continue;
+            }
+
+            let tmpFile = path.resolve(dir, resource);
+            if (fs.existsSync(tmpFile)) {
+              context = dir;
+              resolvedFile = tmpFile;
+              break;
+            }
+            cache.add(dir);
+          }
+        }
       } else {
-        // 2. try to resolve in node modules
-        let res = self.resolveInBuildInfo(buildInfo, request);
+        context = resolveData.context;
+        let tmpFile = path.resolve(context, resource);
+        if (fs.existsSync(tmpFile)) {
+          resolvedFile = tmpFile;
+        }
+      }
+
+      // 2. try to resolve in node modules
+      if (resolvedFile == null) {
+        let res = this.resolveInBuildInfo(buildInfo, resource);
         if (res != null) {
           context = res.context;
           resolvedFile = res.file;
@@ -112,9 +135,10 @@ const UrlDependencyResolver = {
       }
 
       if (resolvedFile != null) {
-        resolveData.request = resolvedFile;
-        dependency.request = resolvedFile;
-        dependency.userRequest = resolvedFile;
+        const resolvedRequest = query ? resolvedFile + '?' + query : resolvedFile;
+        resolveData.request = resolvedRequest;
+        dependency.request = resolvedRequest;
+        dependency.userRequest = resolvedRequest;
         ResourceResolver.addResolvedPath(context);
       }
     }